+++ doc/release-notes.md
@@ -5,6 +5,7 @@
 - upgrade to Postgres 17.2
 - upgrade to Kafka Chart 31.0.0
 - upgrade of example to 3.4.1
+- multi tenancy handling added
 
 # 1.3.5
 - Compatibility with Azure