+++ doc/release-notes.md
@@ -4,8 +4,8 @@
 - upgrade to Prometheus Chart 25.27.0
 - upgrade to Postgres 17.2
 - upgrade to Kafka Chart 31.0.0
-- upgrade of example to 3.4.1
 - multi tenancy handling added
+- Azure Server Handling added 
 
 # 1.3.5
 - Compatibility with Azure